fix for correct usage of number format and refactor

+++ app/src/main/java/projects/hobby/urdufontcomparator/mvp/MainPresenter.java
@@ -94,10 +94,11 @@ public class MainPresenter implements MainMvp.Presenter {
         } else {
             double ratingSum = font.getRatingSum();
             int ratingCount = font.getRatingCount();
-            double ratingAvg = ratingSum / (double) ratingCount;
+            double ratingAvg = ratingSum / ratingCount;
             NumberFormat.getInstance().format(ratingAvg);
             tracker.openFontDetails(font.getName());
-            view.showFontDetailsDialog(font, urduTextSource.prepareFontInfoDialogText(font), ratingAvg,
+            view.showFontDetailsDialog(font, urduTextSource.prepareFontInfoDialogText(font),
+                    ratingAvg,
                     ratingCount);
         }
     }
